test: migrate some more tests from Jest to AVA
