+++ packages/core/src/test/SpyTransport.ts
@@ -1,5 +1,7 @@
 import type { ZWaveLogInfo } from "@zwave-js/core";
 import { ansiRegex, stripColor } from "ansi-colors";
+import type { ExecutionContext } from "ava";
+import sinon from "sinon";
 import { MESSAGE } from "triple-beam";
 import Transport from "winston-transport";
 
@@ -20,10 +22,10 @@ export class SpyTransport extends Transport {
 		super({
 			level: "silly",
 		});
-		this._spy = jest.fn();
+		this._spy = sinon.spy();
 	}
-	private _spy: jest.Mock;
-	public get spy(): jest.Mock {
+	private _spy: sinon.SinonSpy;
+	public get spy(): sinon.SinonSpy {
 		return this._spy;
 	}
 	public log(info: any, next: () => void): any {
@@ -48,8 +50,8 @@ export function assertMessage(
 	}>,
 ): void {
 	const callNumber = options.callNumber || 0;
-	expect(transport.spy.mock.calls.length).toBeGreaterThan(callNumber);
-	const callArg = transport.spy.mock.calls[callNumber][0];
+	expect(transport.spy.callCount).toBeGreaterThan(callNumber);
+	const callArg = transport.spy.getCall(callNumber).args[0];
 	let actualMessage: string = callArg[MESSAGE];
 	// By default ignore the color codes
 	const ignoreColor = options.ignoreColor !== false;
@@ -79,6 +81,54 @@ export function assertMessage(
 	}
 }
 
+/** Tests a printed log message */
+export function assertMessageAva(
+	t: ExecutionContext,
+	transport: SpyTransport,
+	options: Partial<{
+		message: string;
+		predicate: (msg: string) => boolean;
+		/** Default: true */
+		ignoreColor: boolean;
+		/** Default: true */
+		ignoreTimestamp: boolean;
+		/** Default: true */
+		ignoreChannel: boolean;
+		callNumber: number;
+	}>,
+): void {
+	const callNumber = options.callNumber || 0;
+	t.true(transport.spy.callCount > callNumber);
+	const callArg = transport.spy.getCall(callNumber).args[0];
+	let actualMessage: string = callArg[MESSAGE];
+	// By default ignore the color codes
+	const ignoreColor = options.ignoreColor !== false;
+	if (ignoreColor) {
+		actualMessage = stripColor(actualMessage);
+	}
+	// By default, strip away the timestamp and placeholder
+	if (options.ignoreTimestamp !== false) {
+		actualMessage = actualMessage
+			.replace(timestampPrefixRegex, "")
+			.replace(/^ {13}/gm, "");
+	}
+	// by default, strip away the channel label and placeholder
+	if (options.ignoreChannel !== false) {
+		actualMessage = actualMessage
+			.replace(channelPrefixRegex, "")
+			.replace(/^ {7}/gm, "");
+	}
+	if (typeof options.message === "string") {
+		if (ignoreColor) {
+			options.message = stripColor(options.message);
+		}
+		t.is(actualMessage, options.message);
+	}
+	if (typeof options.predicate === "function") {
+		t.true(options.predicate(actualMessage));
+	}
+}
+
 export function assertLogInfo(
 	transport: SpyTransport,
 	options: Partial<{
@@ -88,13 +138,34 @@ export function assertLogInfo(
 	}>,
 ): void {
 	const callNumber = options.callNumber || 0;
-	expect(transport.spy.mock.calls.length).toBeGreaterThan(callNumber);
-	const callArg = transport.spy.mock.calls[callNumber][0];
+	expect(transport.spy.callCount).toBeGreaterThan(callNumber);
+	const callArg = transport.spy.getCall(callNumber).args[0];
 
 	if (typeof options.level === "string") {
 		expect(callArg.level).toEqual(options.level);
 	}
 	if (typeof options.predicate === "function") {
 		expect(callArg).toSatisfy(options.predicate);
+	}
+}
+
+export function assertLogInfoAva(
+	t: ExecutionContext,
+	transport: SpyTransport,
+	options: Partial<{
+		level: string;
+		predicate: (info: ZWaveLogInfo) => boolean;
+		callNumber: number;
+	}>,
+): void {
+	const callNumber = options.callNumber || 0;
+	t.true(transport.spy.callCount > callNumber);
+	const callArg = transport.spy.getCall(callNumber).args[0];
+
+	if (typeof options.level === "string") {
+		t.is(callArg.level, options.level);
+	}
+	if (typeof options.predicate === "function") {
+		t.true(options.predicate(callArg));
 	}
 }
