+++ packages/zwave-js/src/lib/driver/CommandQueueMachine.test.ava.ts
@@ -263,6 +263,7 @@ const testMachine = Machine<
 );
 
 const testModel = createModel<TestContext, TestMachineContext>(
+	// @ts-expect-error
 	testMachine,
 ).withEvents({
 	ADD: {
