+++ packages/zwave-js/tsconfig.build.json
@@ -6,5 +6,5 @@
 		"outDir": "build"
 	},
 	"include": ["src/**/*.ts"],
-	"exclude": ["src/lib/test", "src/**/*.test.ts"]
+	"exclude": ["src/lib/test", "src/**/*.test.ts", "src/**/*.test.ava.ts"]
 }
